App/requirements.txt
altair==5.0.1
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
numpy==1.21.6
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.23.3
pyarrow==12.0.1
pydeck==0.8.1b0
python-dateutil==2.8.2
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0